test & fixed bugs

+++ src/main/java/com/api/soccerTournament/repository/GameRepository.java
@@ -127,7 +127,16 @@ public class GameRepository extends BaseRepository implements IBaseRepository {
 
     @Override
     public Response delete(Integer id) {
-        Response response = dbApi.delete(id, tableName);
+        Response response = readById(id);
+        if (response.statusCode != 0) {
+            return response;
+        }
+
+        if (response.entities.size() == 0) {
+            return new Response(Const.STATUS_CODE_FAIL_GAME_NOT_EXISTS);
+        }
+
+        response = dbApi.delete(id, tableName);
         return response;
     }
 }
